Referential jointing.
Term 'referential joint', not 'waylink'.  The implied meaning
of 'waylink' is too general.  It covers all connections
that involve the way.  It conflicts with other terms.

Terms 'jointer' and 'subjoining', not 'source' and 'target'.
The latter fit poorly a relation whose orientation often varies
in direction (or goes directionless) according to how one views it.

Term hyperlink 'trigger', not 'source'.  'source' fits poorly
a relation whose orientation can vary in direction according
to how one views it - Does the browser move to the new document?
Or the new document to the browser? - Yet retain 'target'
because it accords with the terms of CSS.